example website
hey there all. I went through the first 2 videos and wrote the same code the guy did in the flask videos.

I think for a draft, (the homepage) looked okish. Of course we woud change it lots and make it relevent, these files I made the same as the guy in the video hence they are about a blog (not our app)